+++ groups/item_sites_item_term_stores_item_sets_item_parent_group_sets_request_builder.go
@@ -29,9 +29,9 @@ type ItemSitesItemTermStoresItemSetsItemParentGroupSetsRequestBuilderGetQueryPar
     // Select properties to be returned
     Select []string `uriparametername:"%24select"`
     // Skip the first n items
-    Skip *int32 `uriparametername:"%24skip"`
+    Skip *int64 `uriparametername:"%24skip"`
     // Show only the first n items
-    Top *int32 `uriparametername:"%24top"`
+    Top *int64 `uriparametername:"%24top"`
 }
 // ItemSitesItemTermStoresItemSetsItemParentGroupSetsRequestBuilderGetRequestConfiguration configuration for the request such as headers, query parameters, and middleware options.
 type ItemSitesItemTermStoresItemSetsItemParentGroupSetsRequestBuilderGetRequestConfiguration struct {
