+++ users/item_managed_devices_item_device_compliance_policy_states_request_builder.go
@@ -29,9 +29,9 @@ type ItemManagedDevicesItemDeviceCompliancePolicyStatesRequestBuilderGetQueryPar
     // Select properties to be returned
     Select []string `uriparametername:"%24select"`
     // Skip the first n items
-    Skip *int32 `uriparametername:"%24skip"`
+    Skip *int64 `uriparametername:"%24skip"`
     // Show only the first n items
-    Top *int32 `uriparametername:"%24top"`
+    Top *int64 `uriparametername:"%24top"`
 }
 // ItemManagedDevicesItemDeviceCompliancePolicyStatesRequestBuilderGetRequestConfiguration configuration for the request such as headers, query parameters, and middleware options.
 type ItemManagedDevicesItemDeviceCompliancePolicyStatesRequestBuilderGetRequestConfiguration struct {
